feat: setup dependencies (#1)
* feat: add zod dependency

Add zod dependency for object and data validation.

* feat: add express and prisma as dependencies

This commit adds express and prisma as dependencies to the project. Express is used for creating a web server and prisma is used for connecting to a database.

* feat: add bcrypt for hashing passwords and jest for testing

This commit adds bcrypt for hashing passwords and jest for testing.

This will improve the security of the application by hashing passwords before storing them in the database. Jest will help to ensure that the application is working as expected.

* feat: upgrade babel-jest and add babel-preset-env

This commit upgrades babel-jest to version 29.7.0 and adds babel-preset-env to the project.

The babel-jest upgrade fixes an issue where tests were failing due to an outdated version of babel-jest. The babel-preset-env addition allows for the use of modern JavaScript features in the project.

* feat: add TypeScript support and update jest configs

Add TypeScript support, types for jest, supertest, and bcrypt, and update jest configs to work with TypeScript.

BREAKING CHANGE: jest config has changed, and TypeScript is now required.

* feat: added package-lock.json

* feat: add babel-loader

adds babel-loader to compile typescript to esnext, and added a main typescript file.