feat: added api spec for address management (#4)
* docs: Add Address API spec

Adds a spec for the Address API.

* feat: add create address endpoint

Adds a new "/api/v1/contacts/:idContact/addresses" endpoint for creating addresses.
This endpoint allows users to create new addresses for a specific contact in the system.
The request body should include the following fields:
- street
- city
- province
- country
- postal_code

* docs: add API GET address endpoint

This commit adds an API GET endpoint that retrieves a specific address for a contact.

The new endpoint is available at: GET /api/v1/contacts/`:contactID`/addresses/`:addressID`

* docs: add update address endpoint and request

Adds an endpoint to update an existing address for a contact.

The request body should contain the following fields:

- street
- city
- province
- country
- postal_code

* docs: add remove address endpoint

This commit adds an endpoint to remove an address from a contact.

The endpoint is:

```
DELETE /api/v1/contacts/`:contactID`/addresses/`:addressID`
```

The request header must include:

- `x-api-header`: The API token

The response body will be:

```json
{
  "data": "ok"
}
```

If the address is not found, the response body will be:

```json
{
  "errors": ["address is not found"]
}
```

* docs: add endpoint for listing addresses

Adds an endpoint for listing addresses.

This endpoint allows users to retrieve a list of all addresses associated with a contact.