feat: setup db connection with prisma orm (#5)
* feat: add mysql as a database service

This commit adds a new Docker compose service for MySQL database.

The database service is named `db` and uses the official MySQL image with version 8.0.26 as the base image.
The container is named `ts-restful-db` and uses the port `3306`.
The database is configured with the following environment variables:
- `MYSQL_DATABASE`: The name of the database to be created.
- `MYSQL_USER`: The username of the database user.
- `MYSQL_PASSWORD`: The password of the database user.
- `MYSQL_ROOT_PASSWORD`: The password of the root user.

* feat: configure and connect to MySQL database

Added `.env.example` file for sample database URL
Added `schema.prisma` file to configure and connect to MySQL database

* fix: update database url

Updates the database url in .env.example to point to the correct port.