feat: setup db with prisma orm, setup log with wiston, and setup web server using express  (#9)
* style: remove unnecessary newline

Removes an unnecessary newline from the main.ts file.

* feat: add logging to the application

This commit adds logging to the application using the winston library. The log level isconfigurable via the LOG_LEVEL environment variable.

BREAKING CHANGE: The logging facility is not backwards compatible with the previous version of the application.

* feat: setup webserver

Setup a webserver using express.

* feat(application): add db utilities

Add Prisma Client and configure it to log queries and events.