feat: added layer controller and service for entity user (#10)
* feat: implement generic response and user response models

This commit introduces two new response models:

- GenericResponse: A generic response model that can be used for any type of response.
- UserResponse: A specific response model for user-related responses.

These models provide a consistent and structured way to represent responses from the API, which will make it easier to develop and maintain frontend code.

* feat(error): add error class with caller information

Adds a new `ResponseError` class that includes the caller information (function name, file path, line number, and column number) in the error message.

This is useful for debugging errors, as it provides more context about where the error occurred.

* feat: introduce new code module

Adds a new `Code` module that defines HTTP status codes for use in the application.

* feat: introduce form validation for user register

Introduces a new validation step for the user registration form, ensuring that all required fields are filled out and that the data is in the correct format.

This improves the user experience by preventing them from submitting invalid data and receiving confusing error messages.

* fix(validation): reorder return values

Reorder the return values of the `validate` function to match the expected order of the response.

This change was made to ensure that the response is consistent with the expected order of the response.

* feat: implement user registration

This commit introduces a new feature that allows users to register for an account on the platform.
The registration process involves validating the user's input, checking for existing usernames, hashing the user's password, and storing the user's information in the database.

[BREAKING CHANGE] The user model has been updated to include a hashed password field. This change is necessary to ensure the security of user passwords.

* feat: implement user registration

Implement user registration feature.
This includes creating a new user controller and service, as well as updating the user model.
The registration process includes validating the user's input, checking if the username already exists, and hashing the user's password.

* feat: handle response data and error handling

Handle response data and error handling in user controller.
This ensures a consistent and structured response format, making it easier to consume and handle by the client.

* feat: add public API routes

Added new public API routes for user registration and management.

BREAKING CHANGE: Existing API routes have been refactored to use stricter settings, which may require updates to client code.

* fix: Add error middleware and update response model

This commit fixes an issue where errors were not being properly handled in the application.

- Adds a new error middleware that logs errors and sends a consistent response to the client.
- Updates the response model to use an array of strings for errors, instead of an array of errors.

* feat(middlewares): add middleware for bun

This commit adds a middleware for Bun to handle errors and log them.

The middleware does the following:
- Logs the error stack trace
- Sets the response status to the error code
- Returns the error message in the response body

* fix(error): change caller stack to previous line

This commit changes the caller stack to the previous line in the error class.

This was done to ensure that the caller information is more accurate, as the previous line is more likely to be the actual caller.

* fix(core): add request id when logging error

Add request id when logging error to help debugging

BREAKING CHANGE: the child logger now takes request id as an argument

* feat(users): add register test

Adds a test for the register endpoint. This test checks that the endpoint:

- Returns an error if the username already exists
- Successfully registers a new user