feat: added endpoint to get current user (#13)
* feat: add auth request type

This commit introduces a new `AuthRequest` type to the codebase, which extends the `Request` type and includes an optional `user` property of type `User`. This type is used to represent incoming HTTP requests that may or may not contain user authentication information.

* feat: Add new auth route

Adds a new auth route to the API.
This route requires authentication and will be used to perform user authentication and authorization.

* feat: Middleware for authentication

Added a middleware for authentication that verifies the X-API-TOKEN header and sets the user object on the request if the token is valid.

This middleware helps to protect sensitive routes and ensures that only authorized users can access them.

* feat: implement auth route and middleware error

Implement auth route and middleware error.

Refactor web middleware and route usage.

BREAKING CHANGE: Middleware order has changed.

* feat: implement user controller

Add user controller and corresponding API endpoint to get current user information.

* feat: implement get current user API

Add `getCurrent` endpoint to retrieve the current user.

This endpoint is a protected route and requires the user to be logged in.
If the user is authenticated, the endpoint returns the user's data.
Otherwise, it returns an unauthorized error.

* refactor(users): updated variable names for filename and function

Updated the variable names for filename and function to match the naming convention. This makes the code more consistent and easier to read. Also added validation for token in login response.

* feat: implement web server

This commit implements a simple web server using Bun.

* refactor(tests): convert async functions to arrow functions

This commit converts all async functions in the UserUtil class to arrow functions. This simplifies the code and makes it more concise.

* test(users): add unauthorized check in current user endpoint

Add unauthorized check in current user endpoint to prevent unauthenticated users from accessing user information.

BREAKING CHANGE: Current user endpoint now requires a valid API token for access.