feat: add update current user
This commit adds the ability for a user to update their own information.

The `UserService` now has an `updateCurrent` method, which takes an `UpdateUserRequest` and updates the user's `name` and `password` if they are provided.

The `UpdateUserRequest` type has been added to the `user.model.ts` file.

The `UserValidation` class now has an `UPDATE_CURRENT` type, which is used to validate the `UpdateUserRequest`.