feat: added endpoint to update data current user (#14)
* docs(user): update user example with password field

This commit updates the user example in the docs to include a password field, as the original example was incomplete.

This is a breaking change as it affects the existing user example.

* feat: add update current user

This commit adds the ability for a user to update their own information.

The `UserService` now has an `updateCurrent` method, which takes an `UpdateUserRequest` and updates the user's `name` and `password` if they are provided.

The `UpdateUserRequest` type has been added to the `user.model.ts` file.

The `UserValidation` class now has an `UPDATE_CURRENT` type, which is used to validate the `UpdateUserRequest`.

* feat(user): add update current user endpoint

Add an update current user endpoint to the user controller.

This allows users to update their own information.

* feat: update auth route

Add endpoint to update user

* feat: added test schenario success for update user

* test(users): add test for invalid req body

This commit adds a test to check that the API returns a 400 error when the request body is invalid.

The invalid request body is missing the `name` and `password` fields, which are required for updating a user. The test checks that the API returns the correct error message and that the user is not updated.

* feat(users): add current user authorization

Add test case to ensure that current user update API rejects request with invalid token.