feat: added logout endpoint for user (#15)
* feat: add logoutCurrent method

Add a new method to the UserController class to allow users to log out current.

This change also includes the addition of a new service method to the UserService class to handle the logout process.

* feat(auth): convert logout response to string

Converts the `logout` response in the `user` service to a string to fix the incorrect return type bug.

* feat(users): add logout

Add logout endpoint and tests

This commit adds a new endpoint to the API that allows users to log out. It also includes tests to ensure that the endpoint works as expected.

* feat(users): add logout functionality

Adds a logout functionality to the user model, allowing users to logout of the application.

This includes:
- Adding a new endpoint to the API to handle logout requests
- Updating the user model to track the user's current token
- Adding tests to verify the new functionality