feat(business): change contact association from user to owner
This commit changes the association of a contact from the user who created it to the owner of the contact.

The previous association was causing issues when a user was deleted, as their contacts would also be deleted. This change will prevent this from happening and ensure that contacts are always associated with the correct owner.