feat(error): add custom caller format message
Adds a caller function to the ResponseError class, which allows users to customize the error message. This can be useful for providing more context or information about the error.

BREAKING CHANGE: The error message format has changed. Previous versions of the library relied on the error's file and line properties to generate the error message. This format has been changed to use the caller function instead, which provides more flexibility and control over the error message.