feat: added endpoint to get contact by id (#17)
* feat(business): created models for contact

This commit creates the Contact model and its associated response and create request models.

The Contact model represents the data for a contact, including their first name, last name, email, and phone number.
The ContactResponse model is used to represent the data of a contact when it is returned from the API.
The ContactCreateRequest model is used to represent the data of a contact when it is created in the API.

* feat: add new contact model and service

This commit adds a new Contact model and service to allow users to manage their contacts.
- Add Contact model and validation
- Add ContactService for creating contacts
- Refactor other models to use proper names

* feat(contact): implement create contact feature

This commit introduces a new feature that allows users to create contacts.
It adds a new endpoint to the API that accepts a JSON object with the contact's details, and returns a JSON object with the newly created contact's details.

The endpoint is protected by authentication, and only authenticated users can create contacts.

* feat(contacts): create contact usecase

Introduces a new POST API endpoint under /api/v1/contacts to create a contact.

Adds new tests and migration scripts to support the new endpoint.

* fix: validate contact request

This commit fixes a bug where invalid contact requests were being accepted.

The following validations have been added:
- first_name must have at least 1 character
- phone must have at least 1 character

* test: test must rejected if creating new contact coz unauthorized

test created for unauthorized error while creating new contact

* feat(business): change contact association from user to owner

This commit changes the association of a contact from the user who created it to the owner of the contact.

The previous association was causing issues when a user was deleted, as their contacts would also be deleted. This change will prevent this from happening and ensure that contacts are always associated with the correct owner.

* refactor: Rename prisma client to pc

This change renames the Prisma client from `prismaClient` to `pc` throughout the codebase. The new name is shorter and more concise, and it follows the naming convention used by other popular Node.js libraries.

Additionally, this change updates the Prisma logging configuration to use the `pc` variable instead of `prismaClient`. This ensures that all Prisma-related logs are consistently named and easy to identify.

* feat(error): add custom caller format message

Adds a caller function to the ResponseError class, which allows users to customize the error message. This can be useful for providing more context or information about the error.

BREAKING CHANGE: The error message format has changed. Previous versions of the library relied on the error's file and line properties to generate the error message. This format has been changed to use the caller function instead, which provides more flexibility and control over the error message.

* docs: update endpoint

Update /api/v1/contacts/:id endpoint to use :contactID.

* fix(middleware): improve error handling

Improves error handling by logging the caller's message instead of the file and line number.

* fix(auth): use correct method for contact get route

This commit fixes an issue where the contact get route was using the incorrect method. The correct method is now being used.

* fix(contact service): Async function

Converts toContactResponse to an async function.

* feat(tests): add get endpoint for contacts

Test a new endpoint for getting contacts.

* fix(contact): fix contact not found action for get contact

Added a test case for when the contact is not found. This ensures that the API returns a 404 error when a contact is not found.